+++ http/src/test/scala/com/twitter/finatra/http/tests/server/HttpServerTraitIntegrationTest.scala
@@ -2,12 +2,11 @@ package com.twitter.finatra.http.tests.server
 
 import com.twitter.finagle.Service
 import com.twitter.finagle.http.{Request, Response, Status}
-import com.twitter.finatra.http.EmbeddedHttpServer
-import com.twitter.finatra.http.internal.server.BaseHttpServer
+import com.twitter.finatra.http.{EmbeddedHttpServer, HttpServerTrait}
 import com.twitter.inject.Test
 import com.twitter.util.Future
 
-class FinatraBaseHttpServerIntegrationTest extends Test {
+class HttpServerTraitIntegrationTest extends Test {
 
   test("HiServiceServer#respond") {
     val server = new EmbeddedHttpServer(new HiServer)
@@ -19,11 +18,13 @@ class FinatraBaseHttpServerIntegrationTest extends Test {
   }
 }
 
-class HiServer extends BaseHttpServer {
-  override def httpService = new Service[Request, Response] {
-    def apply(request: Request) = {
+class HiServer extends HttpServerTrait {
+
+  /** Override with an implementation to serve an HTTP Service */
+  override protected def httpService: Service[Request, Response] = new Service[Request, Response] {
+    def apply(request: Request): Future[Response] = {
       val response = Response()
-      response.setStatusCode(200)
+      response.statusCode(200)
       response.setContentString("hi")
       Future(response)
     }